requirements.txt
agno>=1.2.4,<2.0.0
openai==1.78.1
chonkie==1.0.7
googlesearch-python>=1.3.0,<2.0.0
pgvector>=0.4.0,<0.5.0
psycopg[binary,pool]>=3.2.6,<4.0.0
python-dotenv>=1.0.1,<2.0.0
streamlit>=1.44.0,<2.0.0
groq==0.24.0
pycountry==24.6.1
pypdf==5.4.0
SQLAlchemy==2.0.40
sentence-transformers==4.1.0
tensorflow==2.16.2
numpy<2
# pysqlite3-binary
